Ensure that the sample-apps are not edited when reconfiguring.

+++ rstudio/server/start.sh
@@ -7,11 +7,19 @@ echo "Starting RSudio Server"
 
 if [ -n "$SHINY_APPS_PATH" ]; then
 	echo "Updating Shiny server directory path"
-	cp -r /srv/shiny-server/* $SHINY_APPS_PATH
-	rm $SHINY_APPS_PATH/sample-apps
-	rm $SHINY_APPS_PATH/index.html
-	cp -r /opt/shiny-server/samples/sample-apps $SHINY_APPS_PATH/sample-apps
-	cp  /opt/shiny-server/samples/welcome.html $SHINY_APPS_PATH/index.html
+	if [ ! -e $SHINY_APPS_PATH/sample-apps ]; then
+	    cp -r /srv/shiny-server/* $SHINY_APPS_PATH
+	fi
+
+        if [ -L $SHINY_APPS_PATH/sample-apps ]; then
+	    rm -rf $SHINY_APPS_PATH/sample-apps
+	    cp -r /opt/shiny-server/samples/sample-apps $SHINY_APPS_PATH/sample-apps
+	fi
+        if [ -L $SHINY_APPS_PATH/index.html ]; then
+	    rm $SHINY_APPS_PATH/index.html
+	    cp /opt/shiny-server/samples/welcome.html $SHINY_APPS_PATH/index.html
+	fi
+
 	rm -rf /srv/shiny-server
 	ln -s $SHINY_APPS_PATH /srv
 fi
